requirements.txt
click
pydot
pyglet
h5py
pyyaml